+++ netvisor/__init__.py
@@ -7,6 +7,6 @@
     :license: MIT, see LICENSE for more details.
 """
 
-__version__ = '0.3.1'
+__version__ = '0.3.2'
 
 from .core import Netvisor  # flake8: noqa