+++ netvisor/_compat.py
@@ -13,5 +13,7 @@ PY2 = sys.version_info[0] == 2
 
 if not PY2:
     text_type = str
+    string_types = (str,)
 else:
     text_type = unicode
+    string_types = (str, unicode)